Rename StatusBar.tsx to index.tsx; Brute-force fixing file icon not aligning with name issue; Add TSDocs for functions
