+++ new-log-viewer/src/components/MenuBar/index.css
@@ -1,13 +1,13 @@
 .menu-bar {
     display: flex;
-    flexDirection: "row";
+    flex-direction: row;
     height: var(--ylv-status-bar-height);
-    alignItems: "center";
+    align-items: center;
 }
 
 .menu-bar-typography {
     align-items: center;
-    display: flex;
+    display: flex !important;
     flex-grow: 1;
     gap: 2px;
 }
